python/requirements.txt
absl-py==0.3.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bayesian-optimization==1.2.0
bleach==1.5.0
cachetools==2.1.0
certifi==2018.11.29
chardet==2.3.0
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
entrypoints==0.2.3
enum34==1.1.6
gast==0.2.0
google-auth==1.5.1
greenlet==0.4.14
grpcio==1.14.0
h5py==2.8.0
html5lib==0.9999999
idna==2.8
imageio==2.4.1
ipydatawidgets==4.0.0
ipykernel==4.8.2
ipympl==0.2.1
ipython==7.1.1
ipython-genutils==0.2.0
ipyvolume==0.5.1
ipywebrtc==0.4.3
ipywidgets==7.4.2
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-launcher==0.11.2
jupyterlab-server==0.2.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
mock==2.0.0
more-itertools==4.3.0
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.6
networkx==2.1
notebook==5.6.0
numpy==1.15.2
opencv-python==4.0.0.21
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathlib2==2.3.3
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pluggy==0.8.0
prometheus-client==0.3.1
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.2.0
pytest==4.0.1
python-dateutil==2.7.3
pythreejs==2.0.2
pytz==2018.5
PyWavelets==0.5.2
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.21.0
retrying==1.3.3
rsa==3.4.2
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorboard==1.12.0
tensorboardX==1.5
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==1.0.0
torchsummary==1.5.1
torchvision==0.2.1
tornado==5.1
tqdm==4.24.0
traitlets==4.3.2
traittypes==0.2.1
urllib3==1.13.1
virtualenv==15.0.1
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xgboost==0.72.1
